requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dnspython==2.6.1
Flask==3.0.3
google-api-core==2.19.1
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-container==2.50.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
gunicorn==23.0.0
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jira==3.8.0
jmespath==1.0.1
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
soupsieve==2.5
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.4
